actually design a public interface, such as it is
This commit was sponsored by Anika Huhn, Matt Campbell, Devin Prater,
and my other patrons.  If you want to join them, you can support my
work at https://glyph.im/patrons/.
